+++ typing-practice-be/src/main/java/com/typingpractice/typing_practice_be/statistics/scheduler/StatisticsScheduler.java
@@ -1,6 +1,6 @@
 package com.typingpractice.typing_practice_be.statistics.scheduler;
 
-import com.typingpractice.typing_practice_be.statistics.service.StatisticsBatchService;
+import com.typingpractice.typing_practice_be.quote.statistics.service.GlobalQuoteStatisticsBatchService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Scheduled;
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Component;
 @Component
 @RequiredArgsConstructor
 public class StatisticsScheduler {
-  private final StatisticsBatchService batchService;
+  private final GlobalQuoteStatisticsBatchService batchService;
 
   @Scheduled(cron = "0 0 3 * * *")
   public void runDailyBatch() {
